[web] Update InitialReduxStateGate to remove reference to LegacyRawThreadInfo
Summary:
I think @atul missed this one in one of his refactors, and Flow didn't catch it for some reason.

We should not be seeing `LegacyRawThreadInfo` here in the modern `web` client, as `keyserver` should not be sending it to the modern `web` client.

Depends on D12491

Test Plan: Flow

Reviewers: kamil, marcin, atul

Subscribers: