[lib/native] lift createUpdateDBOpsForThreadStoreThreadInfos into lib
Summary:
With migration 77, I had a huge misunderstanding about the new migration approach. I thought for this migration we needed to create and directly return a new thread store state to redux using `.processStoreOperations()`. However, we actually should only just return the operations and just return `state` as is. The `threadStore` will be read by `SQLiteDataHandler` which populates redux with the migrated data from the ops.

Anyways, we need this function for both `native` + `web` migrations, this diff lifts this function into `lib`

Depends on D12497

Test Plan: `flow` + at the end of the stack confirmed that migration 78 works as expected

Reviewers: ashoat, kamil

Reviewed By: ashoat

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12498