[identity] fix wallet account deletion bug
Summary: in D12335 i missed that `get_username_and_password_file` returns an error if the username attribute is missing, causing the whole RPC to fail. rather than modify that db method, i've decided to introduce a new one that just checks for the existence of OPAQUE registration data in a user item. the method returns an error if the lookup fails or if the user doesn't exist.

Resolves https://linear.app/comm/issue/ENG-8474/wallet-account-deletion-not-working

Test Plan: tested on staging -- confirmed that wallet account deletion now works

Reviewers: will!

Subscribers: