[Tunnelbroker] add `MessageToTunnelbrokerRequest` type
Summary:
[ENG-8347](https://linear.app/comm/issue/ENG-8347/add-new-tunnelbroker-message-to-set-device-token)

This is a new message type for messages sent from the device to Tunnelbroker itself.

As a response to that message using `MessageToDeviceRequestStatus` which is exactly what we need - creating follow-up task to rename this to something more generic ([ENG-8609](https://linear.app/comm/issue/ENG-8609/rename-messagetodevicerequeststatus-to-something-more-generic)).

Test Plan: Flow, `cargo test` in `tunnelbroker_messages`

Reviewers: bartek, marcin

Reviewed By: bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12521