[identity] Downgrades search trace error
Summary:
Unsure if this needs to be an error. Addresses https://linear.app/comm/issue/ENG-8463/investigate-identity-search-error-hypererrorparsemethod

Depending on reviewer feedback, if they think this should be kept as an alarm, we can think about filtering down when to trigger the alarm.
Otherwise, we cancel the task

Test Plan:
cargo check

Reviewers:
bartek,
varun

Subscribers: