[web] introduce tagFarcasterChannelByName to CreateFarcasterChannelTagModal
Summary:
This diff introduces the "Other" option in the `CreateFarcasterChannelTagModal` dropdown options. Just like we do in `native` if the user selects this option, we prompt the user with a text input where they can manually input the name of the farcaster channel to create the tag.

I know the large amount of white space does not look super great; however, it is needed so that the size of the modal does not change when the text input is rendered on the modal. Open to other suggestions tho, if anyone has any better ideas

Depends on D12566

Test Plan:
Please see the demo video

{F2101924}

Reviewers: inka, kamil, ashoat

Reviewed By: inka, ashoat

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12567