[lib/web] introduce handleCurrentUserFID to FarcasterDataHandler
Summary: At the moment, the data in the synced metadata table store is not properly synced + shared across different clients. This means that the web client will need to get the fid another way.

Test Plan: Logged out the value of `currentUserFID` from `useCurrentUserFID` and confirmed that the `web` client was getting and returning the correct fid

Reviewers:

Subscribers: