[lib] Wait for connection to auth ks before requesting user infos
Summary:
issue: [[https://linear.app/comm/issue/ENG-8554/login-fails-due-to-race-contidtion | ENG-8554]]
When client receives keyserver auth response, they may call update_relationships endpoint before processing the response. This results in a `not_logged_in` error from the server. To avoid this we will wait until we are connected to the auth keyserver before requesting data. For now all users are required to be connected to the auth keyserver, so this will happen after a while.
We could call `findUserIdentities` and only wait with calling the keyserver, but this would require more changes. And this will all be removed soon, once we stop asking the keyserver for user infos and start asking peers directly. So I think this is all right

Test Plan:
Tested that login doesn't result in errors anymore

Reviewers:
kamil
Subscribers: