[terraform] service connection error alarms
Summary:
This sets up our alarms for errors connecting service A to service B.

We've specified all the relevant log groups in locals. This diff creates alarms that match on `dns error` in logs. The "dns error" error is included as part of tonic when failing to connect

Test Plan: terraform apply. Tested by launching services configured with improper identity domain name on staging and confirming that alarms are activated

Reviewers: bartek, varun

Reviewed By: bartek

Subscribers: tomek, ashoat

Differential Revision: https://phab.comm.dev/D12579