[lib] Change type of `ThinRawThreadInfo.members` from `MemberInfoWithPermissions` to `MemberInfoSansPermissions`
Summary:
Start of stack replacing stack that started with D12469.

This diff makes the change to `ThinRawThreadInfo.members` which intentionally introduces `flow` issues which will be addressed in the rest of the stack.

Also snuck in stripping out MemberInfo.permissions from test data to reduce `flow` noise.

We're at 5 `flow` issues instead of the ~64 we had after D12469. Think this is largely due to stack ending in D12496, but will check each of the places with `flow` issues in previous stack to make sure nothing is broken quietly.

---

Depends on D12554

Test Plan:
NA, `flow`.

This change ONLY changes `flow` types. We still need to update validators, persisted data, `keyserver` endpoint, client migrations, etc. which will be handled in subsequent diffs.

Reviewers: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D12594