[lib] Deprecate `decodeMinimallyEncodedRawThreadInfo`
Summary:
This function is used in 3 places:
1. Unit tests
2. `updateRolesAndPermissions`
3. `deprecatedConvertClientDBThreadInfoToRawThreadInfo`

We will deprecate `updateRolesAndPermissions` in subsequent diff in this stack, at which point this function will only be relevant for keeping existing "legacy" migrations functional. It will not and should NOT be used for migrations going forward because there are no encoded member permissions to be decoded in future persisted `*RawThreadInfo`.

---

Depends on D12597

Test Plan:
This diff is just a rename, will add `flow` suppression in next diff. Looked at all callsites and made sure there's nothing preventing function from being deprecated.

Reviewers:
ashoat