[lib] Deprecate `updateRolesAndPermissions`
Summary:
We want to deprecate this function because it's no longer relevant for `*RawThreadInfo`s without member permissions.

We're keeping it around to ensure existing migrations work as expected.

---

Depends on D12598

Test Plan: This diff is just a rename to make it clear function shouldn't be used in future. Looked at all callsites and ensured there's nothing preventing function from being deprecated.

Reviewers: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D12599