[terraform] add APNs secrets config
Summary: This will be added to AWS KMS secret manager as JSON and is exposed as an environment variable in the container.

Test Plan:
`yarn init-local-cloud` - the exact I'll test exact functionality on staging after entire stack will be accepted.
On local setup this config will be passed using CMD args.

Reviewers: bartek, marcin

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12606