[landing] extend exiting competitor comparison code to handle new differentiation section
Summary: This diff extends our preexiting logic for the competitor comparison to better handle a new section explaining how Comm is differentiated. At the moment, the copy is not ready and given that I will be leaving in a few days, I made the copy for this new section just be lorum ipsum dummy text that should be easy to update and replace.

Test Plan:
`flow` + confirmed later on in the stack that the new section was properly introduced

{F2159541}

Reviewers: inka, ashoat

Reviewed By: ashoat

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12629