[keyserver][lib][native] Continue storing password hash in authoritative keyserver MariaDB
Summary:
We want to make sure clients from before the rollout can still auth. We'll revert this soon after launch, once all of the clients have been confirmed to be updated. We don't need to keep it around longer than that because this only concerned brand new users, and brand new users can be expected to have the most recent version of the app.

This addresses [ENG-8692](https://linear.app/comm/issue/ENG-8692/continue-storing-password-hash-in-authoritative-keyserver-mariadb).

Test Plan:
1. I created a password account on an iOS simulator with usingCSAT=true
2. I set usingCSAT=false
3. I tried logging in on web and confirmed it worked

Reviewers: tomek, inka

Reviewed By: inka

Differential Revision: https://phab.comm.dev/D12641