[lib] Introduce DMOperations
Summary:
This is the start of a stack addressing [ENG-8421](https://linear.app/comm/issue/ENG-8421/introduce-type-for-e2ee-dm-updates).

The idea is that communication will occur between peer devices via this new message type, communicated across `usePeerToPeerMessageHandler` via Tunnelbroker.

We might also process these on the sending client.

Test Plan: I haven't tested this stack outside of Flow. I'd like to propose that we land this so the team can iterate. I'll make sure to test all of the functionality once it's integrated into the rest of the codebase.

Reviewers: tomek, marcin, inka

Reviewed By: inka

Differential Revision: https://phab.comm.dev/D12652