[lib] DMOperationSpec for SEND_TEXT_MESSAGE operation
Summary: I did this one first because it's more simple

Depends on D12652

Test Plan: I haven't tested this stack outside of Flow. I'd like to propose that we land this so the team can iterate. I'll make sure to test all of the functionality once it's integrated into the rest of the codebase.

Reviewers: tomek, marcin, inka

Subscribers: