[lib] DMOperationSpec for CREATE_THREAD operation
Summary:
Handles creating all `ThickThreadType`s, including sidebars.

The `RawThreadInfo` construction was loosely inspired by `createPendingThread`, but it's different because pending threads are intentionally constructed to have limited permissions, so eg. the user can't go the settings of a pending thread.

Depends on D12657

Test Plan: I haven't tested this stack outside of Flow. I'd like to propose that we land this so the team can iterate. I'll make sure to test all of the functionality once it's integrated into the rest of the codebase.

Reviewers: tomek, marcin, inka

Differential Revision: https://phab.comm.dev/D12658