[lib] Process DMOperations when received from peers via Tunnelbroker
Summary: This is part of addressing [ENG-8425](https://linear.app/comm/issue/ENG-8425/update-usepeertopeermessagehandler-to-handle-e2ee-dm-updates).

I'll submit a follow-up diff that introduces a new action here to process the resultant data we receive.

Submitting this now so @tomek can see the code, as his work relates to this code.

Depends on D12740

Test Plan: I haven't had a chance to really test this yet.

Reviewers: tomek, kamil, marcin

Subscribers: