[lib] Fix crash when pressing message from anonymous author
Summary:
If we don't have a `username` for a message author, the app currently crashing when you press the message.

It looks like the Flow types here are unnecessarily strict. Once I relaxed them, I was able to remove the `invariant`.

This resolves [ENG-8777](https://linear.app/comm/issue/ENG-8777/dont-crash-when-pressing-message-with-missing-author-username).

Depends on D12679

Test Plan: I reverted D12674 to reproduce the issue. I then created a new user from a separate account, and sent a message to my main user. The main user was able to press the message, react, and start a sidebar on the messages for the other user, even though the other user showed up as "anonymous"

Reviewers: inka, tomek

Differential Revision: https://phab.comm.dev/D12680