[native] update connect farcaster prompt to be less annyoying
Summary: We want to update the logic for when we show the farcaster prompt so that it is less annoying. This means we only want to show this prompt to the user just once if the user has an existing account that does not have an fid yet. We will also not show this prompt to users who have created a new account and have declined to connect their farcaster account in the registaration flow

Test Plan: Confirmed that the following 3 scenarios work as expected

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12683