[native] update connect farcaster prompt to be less annyoying
Summary:
We want to update the logic for when we show the farcaster prompt so that it is less annoying. This means we only want to show this prompt to the user just once if the user has an existing account that does not have an fid yet. We will also avoid showing the prompt to users who created new accounts and chose not to connect their farcaster account during the registration process

Linear task: https://linear.app/comm/issue/ENG-8449/rethink-farcaster-prompt-to-make-sure-its-not-annoying

Test Plan:
Confirmed that the following 3 scenarios work as expected

1. Existing user with `fid` => No prompt
2. Existing user with no `fid` => Prompt
3. New user who declined to connect to farcaster => No prompot

Reviewers: ashoat, inka

Reviewed By: ashoat

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12683