[lib] Export once at the end of DMOperationSpec files
Summary:
In the last diff I introduced a second export for `create-thread-spec.js`, and I thought it was confusing that the exports were listed individually. I decided to move to a single export at the end of the file, but am splitting it out into its own diff since I wanted to update the other `DMOperationSpec`s too for consistency.

Depends on D12658

Test Plan: Flow

Reviewers: tomek, marcin, inka

Reviewed By: tomek

Differential Revision: https://phab.comm.dev/D12699