[terraform] add authoritative keyserver id as a user-specified variable
Summary:
Introduce specifying the authoritative keyserver id as a terraform variable which translates to a keyserver environmental variable

Depends on D12709

Test Plan:
In our prior testing, wasn't able to connect a debug native client without specifying the authoritative id to the keyserver.
With this configuration, was able to connect properly. Previously set this manually

Reviewers: varun, bartek

Reviewed By: bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12710