[lib] Add `invariant` to suppress `flow` issue in `getPushUserInfo`
Summary:
There are new client-side member permission checks in `getPushUserInfo` which I believe (and hope...) are only relevant for `ThickRawThreadInfo`s.

`getPushUserInfo` gets called by `preparePushNotifs` gets called by `usePreparePushNotifs` which isn't used anywhere at this point.

Putting in `invariant`s instead of `$FlowIgnore` so that it's harder to miss if any `ThinRawThreadInfo`s make it to this point... which should be an error case because `ThinRawThreadInfo`s will not have any permissions.

Fixing things here is tracked in https://linear.app/comm/issue/ENG-8809/modify-functions-in-send-utilsjs-in-to-operate-only-on

---

Depends on D12597

Test Plan: `flow`

Reviewers: ashoat, marcin

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D12722