[keyserver] include keyserver url in the farcaster channel tag blob
Summary: When we add + auth keyservers that are unknown we will need to make sure the keyserver url we are trying to auth + add is valid. To do this with the auto join community handler we will need to include the keyserver url in the farcaster channel tag blob payload so that we can validate it later.

Test Plan: Created a farcaster channel tag blob and logged out + confirmed that the keyserver url was present in the blob payload

Reviewers: ashoat, inka

Reviewed By: ashoat

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12702