[native] update AutoJoinCommunityHandler to use useJoinCommunity hook
Summary:
This diff updates `AutoJoinCommunityHandler` to use the `useJoinCommunity` hook. Since the logic for `useJoinCommunity` can only handle one community at a time, we needed to create a secondary handler component (`JoinHandler`) which handles calling `useJoinCommunity` and then awaiting the promise that gets returned by the hook.

Depends on D12713

Test Plan: Confirmed that I was still able to auto join any community that was tagged to a farcaster channel that I am following on farcaster

Reviewers: ashoat, inka

Reviewed By: ashoat

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12714