[web] Don't allow creating subchannels with users the user should not be able to add
Summary:
issue: [[https://linear.app/comm/issue/ENG-8710/creating-a-subchannel-doesnt-work-in-genesis-subchannels | ENG-8710]]
The user should not be able to add some users to some chats. On native we handle this in the UI by showing a popup if the user wants to add someone they cannot.
On web we decided that for now we will just not display those users. Handlig tih siwth a popup is tracked in [[https://linear.app/comm/issue/ENG-8798/adding-non-friends-to-genesis-subchannels | ENG-8798]]

Property `alert` is set only when the user cannot be added to the thread. There are four cases. They can be found in lib/shared/search-utils.js in `usePotentialMemberItems`.

Test Plan: Tested that in all four cses the users don't show up in search results on web anymore.

Reviewers: tomek, kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D12715