[lib/native] lift auto join community handler to lib
Summary:
We are going to want to use the auto join community handler in web. Before we can do that though, we need to lift the auto join community handler into `lib`. One thing to note is that there is a difference in `calendarQuery` between `web` and `native`, so we need to separate this logic for each platform.

Depends on D12714

Test Plan: `flow` + confirmed there were no regressions

Reviewers: ashoat, inka

Reviewed By: ashoat

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12716