[lib] Fix subchannels of subchannels in GENESIS not appearing in navigation side drawer
Summary: This occurred because I misused the `ChatMentionContext`. We should use `getCommunityThreadIDForGenesisThreads` instead of `getCommunity`, so that channels of GENESIS are treated the same as community roots.

This addresses [ENG-8808](https://linear.app/comm/issue/ENG-8808/community-drawer-ui-is-broken).

Test Plan: I reproduced the issue locally and confirmed that this diff resolves it

Reviewers: inka, tomek

Subscribers: