[terraform] Make number of secondary nodes configured by the user
Summary:
This allows the user to configure the amount of secondary nodes they want to launch.

I think that we'll want to consider some more advanced auto-scaling logic in the future to automatically launch secondary nodes, but configuring this to be
user-defined for now

Depends on D12725

Test Plan:
terraform apply. Setting variable to 2 caused 2 secondary nodes tasks to deploy

Reviewers:
varun,
bartek

Subscribers: