[lib] add device type to qr code URL
Summary:
we need to know whether the device being scanned is a keyserver so that we don't accidentally add multiple keyservers to a user's device list.

this change is backwards compatible. if a device presents a QR code without a device type, it will still be parsed successfully.

Test Plan:
- scanned QR code with old URL (sans device type) successfully
- scanned QR code with new URL successfully
- scanned QR code with new URL successfully from old device unaware of new URL format

Reviewers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12734