[Tunnelbroker] implement OAuth 2.0 JWT token encoding
Summary:
Implementing what is described in Google's OAuth 2.0 for Server to Server Applications, [Forming the JWT claim set
](https://developers.google.com/identity/protocols/oauth2/service-account#authorizingrequests).

`get_auth_bearer` is only public method, will handle generating and refreshing token.

Depends on D12753

Test Plan:
Call `get_jwt_token` and test if returns the correct JWT token in the format
```
{Base64url encoded header}.{Base64url encoded claim set}
```
Later in the stack this token is used to make an HTTP call to get an access token for the FCM push service.

Reviewers: bartek, marcin

Reviewed By: bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12754