[lib] Move `stripPermissionsFromMemberInfo` out of `...-types.js` file
Summary:
We're going to create some "higher order" utility functions using `stripPermissionsFromMemberInfo`, so figured it would make sense to pull out of `types` file and into its own utility file.

For context, we're going to want to mirror the utility files in `special-roles.js` where we can patch `RawThreadInfo` and `RawThreadInfos` all in one go. This will make the changes to `rawThreadInfoFromServerThreadInfo` and the client migrations easier.

---

Depends on D12755

Test Plan: This is just a move diff, unit tests still pass. Note: There's still a validator unit test that is expected to fail. Going to address that now that `flow` stuff is reolved.

Reviewers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D12756