[native] Extract MessageType to DatabaseManagers
Summary:
I want to reuse this from the `DatabaseManagers` codebase, which is isolated from `NativeModules` so that it can be compiled to WASM for `web`.

As such, I'm moving it into that folder, where it can be accessed from the whole codebase.

Depends on D12760

Test Plan: Make sure it compiles

Reviewers: kamil, tomek, marcin

Subscribers: