[lib] Introduce new PROCESS_DM_OPS action and dispatch from useProcessDMOperation
Summary:
Reducers will be handled in the next diff; this diff just introduces the action, and dispatches it.

Submitting this now so @tomek can see the code, as his work relates to this code.

Depends on D12741

Test Plan: Test plan described [here](https://linear.app/comm/issue/ENG-8426/make-sure-rawthreadinfos-with-ids-that-dont-have-or-character-dont#comment-f32630d3)

Reviewers: tomek, kamil, marcin

Reviewed By: tomek

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D12780